% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{create_word_dict}
\alias{create_word_dict}
\title{create_word_dict}
\usage{
create_word_dict(lexicon)
}
\arguments{
\item{lexicon}{a named list of vectors.}
}
\value{
Dictionary
}
\description{
create_word_dict
}
