% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoder.R
\docType{data}
\name{CriterionTypes}
\alias{CriterionTypes}
\title{CriterionTypes}
\format{
## `CriterionTypes`
A list with 3 items:
\describe{
  \item{ASG}{string "ASG"}
  \item{CTC}{string "CTC"}
  \item{S2S}{string "S2S"}
}
}
\usage{
CriterionTypes
}
\description{
A enum (list) of criterion types.
}
\keyword{datasets}
