% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoder.R
\name{Decoder}
\alias{Decoder}
\title{Decoder}
\description{
Decoder

Decoder
}
\examples{

library(flashlighttext)

decoder <- Decoder$new()
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ptr}}{set and get the pointer to a Decoder instance.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Decoder-new}{\code{Decoder$new()}}
\item \href{#method-Decoder-decode_begin}{\code{Decoder$decode_begin()}}
\item \href{#method-Decoder-decode_step}{\code{Decoder$decode_step()}}
\item \href{#method-Decoder-decode_end}{\code{Decoder$decode_end()}}
\item \href{#method-Decoder-decode}{\code{Decoder$decode()}}
\item \href{#method-Decoder-prune}{\code{Decoder$prune()}}
\item \href{#method-Decoder-n_decoded_frames_in_buffer}{\code{Decoder$n_decoded_frames_in_buffer()}}
\item \href{#method-Decoder-get_best_hypothesis}{\code{Decoder$get_best_hypothesis()}}
\item \href{#method-Decoder-get_all_final_hypothesis}{\code{Decoder$get_all_final_hypothesis()}}
\item \href{#method-Decoder-clone}{\code{Decoder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Decoder-new"></a>}}
\if{latex}{\out{\hypertarget{method-Decoder-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Decoder$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Decoder-decode_begin"></a>}}
\if{latex}{\out{\hypertarget{method-Decoder-decode_begin}{}}}
\subsection{Method \code{decode_begin()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Decoder$decode_begin()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Decoder-decode_step"></a>}}
\if{latex}{\out{\hypertarget{method-Decoder-decode_step}{}}}
\subsection{Method \code{decode_step()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Decoder$decode_step(emissions, T, N)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{emissions}}{a emissions}

\item{\code{T}}{a T}

\item{\code{N}}{a N}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Decoder-decode_end"></a>}}
\if{latex}{\out{\hypertarget{method-Decoder-decode_end}{}}}
\subsection{Method \code{decode_end()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Decoder$decode_end()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Decoder-decode"></a>}}
\if{latex}{\out{\hypertarget{method-Decoder-decode}{}}}
\subsection{Method \code{decode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Decoder$decode(emissions, T, N)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{emissions}}{a emissions}

\item{\code{T}}{a T}

\item{\code{N}}{a N}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Decoder-prune"></a>}}
\if{latex}{\out{\hypertarget{method-Decoder-prune}{}}}
\subsection{Method \code{prune()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Decoder$prune(lookBack = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lookBack}}{a lookBack}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Decoder-n_decoded_frames_in_buffer"></a>}}
\if{latex}{\out{\hypertarget{method-Decoder-n_decoded_frames_in_buffer}{}}}
\subsection{Method \code{n_decoded_frames_in_buffer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Decoder$n_decoded_frames_in_buffer()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Decoder-get_best_hypothesis"></a>}}
\if{latex}{\out{\hypertarget{method-Decoder-get_best_hypothesis}{}}}
\subsection{Method \code{get_best_hypothesis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Decoder$get_best_hypothesis(lookBack = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lookBack}}{a lookBack}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Decoder-get_all_final_hypothesis"></a>}}
\if{latex}{\out{\hypertarget{method-Decoder-get_all_final_hypothesis}{}}}
\subsection{Method \code{get_all_final_hypothesis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Decoder$get_all_final_hypothesis()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Decoder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Decoder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Decoder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
