% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{Dictionary}
\alias{Dictionary}
\title{Dictionary}
\value{
Dictionary
}
\description{
Dictionary

Dictionary
}
\examples{

## ------------------------------------------------
## Method `Dictionary$new`
## ------------------------------------------------

library(flashlighttext)

dict <- Dictionary$new(letters)
dict$get_entry(0)
dict$get_index("b")
dict$map_indices_to_entries(1:3)
dict$entry_size()
dict$add_entry("A")
dict$contains("A")
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ptr}}{has a single parameter new_ptr that accepts a new pointer to a Dictionary.
It returns invisible(NULL)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Dictionary-new}{\code{Dictionary$new()}}
\item \href{#method-Dictionary-entry_size}{\code{Dictionary$entry_size()}}
\item \href{#method-Dictionary-index_size}{\code{Dictionary$index_size()}}
\item \href{#method-Dictionary-add_entry}{\code{Dictionary$add_entry()}}
\item \href{#method-Dictionary-get_index}{\code{Dictionary$get_index()}}
\item \href{#method-Dictionary-get_entry}{\code{Dictionary$get_entry()}}
\item \href{#method-Dictionary-contains}{\code{Dictionary$contains()}}
\item \href{#method-Dictionary-map_indices_to_entries}{\code{Dictionary$map_indices_to_entries()}}
\item \href{#method-Dictionary-map_entries_to_indices}{\code{Dictionary$map_entries_to_indices()}}
\item \href{#method-Dictionary-is_contiguous}{\code{Dictionary$is_contiguous()}}
\item \href{#method-Dictionary-set_default_index}{\code{Dictionary$set_default_index()}}
\item \href{#method-Dictionary-clone}{\code{Dictionary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dictionary-new"></a>}}
\if{latex}{\out{\hypertarget{method-Dictionary-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dictionary$new(tokens = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tokens}}{file or list containing valid tokens. If using a file, the 
expected format is for tokens mapping to the same index to be on the same
line}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(flashlighttext)

dict <- Dictionary$new(letters)
dict$get_entry(0)
dict$get_index("b")
dict$map_indices_to_entries(1:3)
dict$entry_size()
dict$add_entry("A")
dict$contains("A")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dictionary-entry_size"></a>}}
\if{latex}{\out{\hypertarget{method-Dictionary-entry_size}{}}}
\subsection{Method \code{entry_size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dictionary$entry_size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dictionary-index_size"></a>}}
\if{latex}{\out{\hypertarget{method-Dictionary-index_size}{}}}
\subsection{Method \code{index_size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dictionary$index_size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dictionary-add_entry"></a>}}
\if{latex}{\out{\hypertarget{method-Dictionary-add_entry}{}}}
\subsection{Method \code{add_entry()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dictionary$add_entry(entry, idx = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entry}}{an entry}

\item{\code{idx}}{an idx}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dictionary-get_index"></a>}}
\if{latex}{\out{\hypertarget{method-Dictionary-get_index}{}}}
\subsection{Method \code{get_index()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dictionary$get_index(entry)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entry}}{an entry}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the index
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dictionary-get_entry"></a>}}
\if{latex}{\out{\hypertarget{method-Dictionary-get_entry}{}}}
\subsection{Method \code{get_entry()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dictionary$get_entry(idx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{idx}}{an index}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the entry
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dictionary-contains"></a>}}
\if{latex}{\out{\hypertarget{method-Dictionary-contains}{}}}
\subsection{Method \code{contains()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dictionary$contains(entry)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entry}}{an entry}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a boolean
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dictionary-map_indices_to_entries"></a>}}
\if{latex}{\out{\hypertarget{method-Dictionary-map_indices_to_entries}{}}}
\subsection{Method \code{map_indices_to_entries()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dictionary$map_indices_to_entries(indices)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{indices}}{a vector of indices}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dictionary-map_entries_to_indices"></a>}}
\if{latex}{\out{\hypertarget{method-Dictionary-map_entries_to_indices}{}}}
\subsection{Method \code{map_entries_to_indices()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dictionary$map_entries_to_indices(entries)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entries}}{a vector of entries}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dictionary-is_contiguous"></a>}}
\if{latex}{\out{\hypertarget{method-Dictionary-is_contiguous}{}}}
\subsection{Method \code{is_contiguous()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dictionary$is_contiguous()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a boolean
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dictionary-set_default_index"></a>}}
\if{latex}{\out{\hypertarget{method-Dictionary-set_default_index}{}}}
\subsection{Method \code{set_default_index()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dictionary$set_default_index(idx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{idx}}{one index}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dictionary-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Dictionary-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dictionary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
