% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f}
\alias{f}
\title{Format and interpolate a string}
\usage{
f(
  ...,
  .sep = "",
  .envir = parent.frame(),
  .open = "{",
  .close = "}",
  .na = "NA",
  .null = character(),
  .comment = "#",
  .literal = FALSE,
  .transformer = identity_transformer,
  .trim = TRUE
)
}
\description{
See \code{glue::glue} for details.
}
\examples{
f("iris dataset has {nrow(iris)} rows and {ncol(iris)} columns.")

}
