% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenlm.R
\name{KenLM}
\alias{KenLM}
\title{KenLM}
\description{
KenLM

KenLM
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{filename}}{returns a string}

\item{\code{tokens}}{returns a string vector}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ptr}}{has a single parameter new_ptr that accepts a new pointer to a KenLM.
It returns invisible(NULL)}

\item{\code{filename}}{returns a string}

\item{\code{tokens}}{returns a string vector}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KenLM-new}{\code{KenLM$new()}}
\item \href{#method-KenLM-clone}{\code{KenLM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KenLM-new"></a>}}
\if{latex}{\out{\hypertarget{method-KenLM-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KenLM$new(path = NULL, usrTknDict = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KenLM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KenLM-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KenLM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
