% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon_free_decoder.R
\name{LexiconFreeDecoder}
\alias{LexiconFreeDecoder}
\title{LexiconFreeDecoder}
\description{
LexiconFreeDecoder

LexiconFreeDecoder
}
\section{Super class}{
\code{\link[flashlighttext:Decoder]{flashlighttext::Decoder}} -> \code{LexiconFreeDecoder}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ptr}}{set and get the pointer to a LexiconFreeDecoder instance.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LexiconFreeDecoder-new}{\code{LexiconFreeDecoder$new()}}
\item \href{#method-LexiconFreeDecoder-decode_begin}{\code{LexiconFreeDecoder$decode_begin()}}
\item \href{#method-LexiconFreeDecoder-decode_step}{\code{LexiconFreeDecoder$decode_step()}}
\item \href{#method-LexiconFreeDecoder-decode_end}{\code{LexiconFreeDecoder$decode_end()}}
\item \href{#method-LexiconFreeDecoder-decode}{\code{LexiconFreeDecoder$decode()}}
\item \href{#method-LexiconFreeDecoder-n_hypothesis}{\code{LexiconFreeDecoder$n_hypothesis()}}
\item \href{#method-LexiconFreeDecoder-prune}{\code{LexiconFreeDecoder$prune()}}
\item \href{#method-LexiconFreeDecoder-n_decoded_frames_in_buffer}{\code{LexiconFreeDecoder$n_decoded_frames_in_buffer()}}
\item \href{#method-LexiconFreeDecoder-get_best_hypothesis}{\code{LexiconFreeDecoder$get_best_hypothesis()}}
\item \href{#method-LexiconFreeDecoder-get_all_final_hypothesis}{\code{LexiconFreeDecoder$get_all_final_hypothesis()}}
\item \href{#method-LexiconFreeDecoder-get_transitions}{\code{LexiconFreeDecoder$get_transitions()}}
\item \href{#method-LexiconFreeDecoder-get_options}{\code{LexiconFreeDecoder$get_options()}}
\item \href{#method-LexiconFreeDecoder-get_blank_idx}{\code{LexiconFreeDecoder$get_blank_idx()}}
\item \href{#method-LexiconFreeDecoder-get_sil_idx}{\code{LexiconFreeDecoder$get_sil_idx()}}
\item \href{#method-LexiconFreeDecoder-clone}{\code{LexiconFreeDecoder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoder-new"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoder-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoder$new(
  options,
  lm,
  sil_token_idx,
  blank_token_idx,
  transitions
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{options}}{a options}

\item{\code{lm}}{a lm}

\item{\code{sil_token_idx}}{a sil_token_idx}

\item{\code{blank_token_idx}}{a blank_token_idx}

\item{\code{transitions}}{a transitions}

\item{\code{is_lm_token}}{a is_lm_token}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
LexiconFreeDecoder
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoder-decode_begin"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoder-decode_begin}{}}}
\subsection{Method \code{decode_begin()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoder$decode_begin()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoder-decode_step"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoder-decode_step}{}}}
\subsection{Method \code{decode_step()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoder$decode_step(emissions, T, N)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{emissions}}{a emissions}

\item{\code{T}}{a T}

\item{\code{N}}{a N}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoder-decode_end"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoder-decode_end}{}}}
\subsection{Method \code{decode_end()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoder$decode_end()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoder-decode"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoder-decode}{}}}
\subsection{Method \code{decode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoder$decode(emissions, T, N)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{emissions}}{a emissions}

\item{\code{T}}{a T}

\item{\code{N}}{a N}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoder-n_hypothesis"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoder-n_hypothesis}{}}}
\subsection{Method \code{n_hypothesis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoder$n_hypothesis()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
int
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoder-prune"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoder-prune}{}}}
\subsection{Method \code{prune()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoder$prune(lookBack = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lookBack}}{a lookBack}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoder-n_decoded_frames_in_buffer"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoder-n_decoded_frames_in_buffer}{}}}
\subsection{Method \code{n_decoded_frames_in_buffer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoder$n_decoded_frames_in_buffer()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
int
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoder-get_best_hypothesis"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoder-get_best_hypothesis}{}}}
\subsection{Method \code{get_best_hypothesis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoder$get_best_hypothesis(lookBack = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lookBack}}{a lookBack}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
DecodeResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoder-get_all_final_hypothesis"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoder-get_all_final_hypothesis}{}}}
\subsection{Method \code{get_all_final_hypothesis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoder$get_all_final_hypothesis()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of DecodeResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoder-get_transitions"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoder-get_transitions}{}}}
\subsection{Method \code{get_transitions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoder$get_transitions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of DecodeResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoder-get_options"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoder-get_options}{}}}
\subsection{Method \code{get_options()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoder$get_options()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of DecodeResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoder-get_blank_idx"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoder-get_blank_idx}{}}}
\subsection{Method \code{get_blank_idx()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoder$get_blank_idx()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of DecodeResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoder-get_sil_idx"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoder-get_sil_idx}{}}}
\subsection{Method \code{get_sil_idx()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoder$get_sil_idx()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of DecodeResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
