% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon_free_decoder.R
\name{LexiconFreeDecoderOptions}
\alias{LexiconFreeDecoderOptions}
\title{LexiconFreeDecoderOptions}
\value{
LexiconFreeDecoderOptions
}
\description{
LexiconFreeDecoderOptions

LexiconFreeDecoderOptions
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{beam_size}}{Maximum number of hypothesis we hold after each step}

\item{\code{beam_size_token}}{Maximum number of tokens we consider at each step}

\item{\code{beam_threshold}}{Threshold to prune hypothesis}

\item{\code{lm_weight}}{Weight of lm}

\item{\code{sil_score}}{Silence insertion score}

\item{\code{log_add}}{If or not use logadd when merging hypothesis}

\item{\code{criterion_type}}{a CriterionType: "CTC" or "ASG" (see [CriterionTypes])}

\item{\code{ptr}}{set and get the pointer to a LexiconFreeDecoderOptions instance.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LexiconFreeDecoderOptions-new}{\code{LexiconFreeDecoderOptions$new()}}
\item \href{#method-LexiconFreeDecoderOptions-clone}{\code{LexiconFreeDecoderOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoderOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoderOptions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoderOptions$new(
  beam_size,
  beam_size_token,
  beam_threshold,
  lm_weight,
  sil_score,
  log_add,
  criterion_type
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{beam_size}}{Maximum number of hypothesis we hold after each step}

\item{\code{beam_size_token}}{Maximum number of tokens we consider at each step}

\item{\code{beam_threshold}}{Threshold to prune hypothesis}

\item{\code{lm_weight}}{Weight of lm}

\item{\code{sil_score}}{Silence insertion score}

\item{\code{log_add}}{If or not use logadd when merging hypothesis}

\item{\code{criterion_type}}{a CriterionType: "CTC" or "ASG" (see [CriterionTypes])}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoderOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoderOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoderOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
