% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon_free_seq2seq_decoder.R
\name{LexiconFreeSeq2SeqDecoder}
\alias{LexiconFreeSeq2SeqDecoder}
\title{LexiconFreeSeq2SeqDecoder}
\description{
LexiconFreeSeq2SeqDecoder

LexiconFreeSeq2SeqDecoder
}
\section{Super class}{
\code{\link[flashlighttext:Decoder]{flashlighttext::Decoder}} -> \code{LexiconFreeSeq2SeqDecoder}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ptr}}{set and get the pointer to a LexiconFreeSeq2SeqDecoder instance.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LexiconFreeSeq2SeqDecoder-new}{\code{LexiconFreeSeq2SeqDecoder$new()}}
\item \href{#method-LexiconFreeSeq2SeqDecoder-decode_begin}{\code{LexiconFreeSeq2SeqDecoder$decode_begin()}}
\item \href{#method-LexiconFreeSeq2SeqDecoder-decode_step}{\code{LexiconFreeSeq2SeqDecoder$decode_step()}}
\item \href{#method-LexiconFreeSeq2SeqDecoder-decode_end}{\code{LexiconFreeSeq2SeqDecoder$decode_end()}}
\item \href{#method-LexiconFreeSeq2SeqDecoder-decode}{\code{LexiconFreeSeq2SeqDecoder$decode()}}
\item \href{#method-LexiconFreeSeq2SeqDecoder-n_hypothesis}{\code{LexiconFreeSeq2SeqDecoder$n_hypothesis()}}
\item \href{#method-LexiconFreeSeq2SeqDecoder-prune}{\code{LexiconFreeSeq2SeqDecoder$prune()}}
\item \href{#method-LexiconFreeSeq2SeqDecoder-n_decoded_frames_in_buffer}{\code{LexiconFreeSeq2SeqDecoder$n_decoded_frames_in_buffer()}}
\item \href{#method-LexiconFreeSeq2SeqDecoder-get_best_hypothesis}{\code{LexiconFreeSeq2SeqDecoder$get_best_hypothesis()}}
\item \href{#method-LexiconFreeSeq2SeqDecoder-get_all_final_hypothesis}{\code{LexiconFreeSeq2SeqDecoder$get_all_final_hypothesis()}}
\item \href{#method-LexiconFreeSeq2SeqDecoder-clone}{\code{LexiconFreeSeq2SeqDecoder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeSeq2SeqDecoder-new"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeSeq2SeqDecoder-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeSeq2SeqDecoder$new(
  options,
  lm,
  eos_token_idx,
  emitting_model_update_func,
  max_output_length
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{options}}{an options}

\item{\code{lm}}{a lm}

\item{\code{eos_token_idx}}{an eos_token_idx}

\item{\code{emitting_model_update_func}}{an emitting_model_update_func}

\item{\code{max_output_length}}{a max_output_length}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
LexiconFreeSeq2SeqDecoder
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeSeq2SeqDecoder-decode_begin"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeSeq2SeqDecoder-decode_begin}{}}}
\subsection{Method \code{decode_begin()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeSeq2SeqDecoder$decode_begin()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeSeq2SeqDecoder-decode_step"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeSeq2SeqDecoder-decode_step}{}}}
\subsection{Method \code{decode_step()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeSeq2SeqDecoder$decode_step(emissions, T, N)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{emissions}}{a emissions}

\item{\code{T}}{a T}

\item{\code{N}}{a N}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeSeq2SeqDecoder-decode_end"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeSeq2SeqDecoder-decode_end}{}}}
\subsection{Method \code{decode_end()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeSeq2SeqDecoder$decode_end()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeSeq2SeqDecoder-decode"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeSeq2SeqDecoder-decode}{}}}
\subsection{Method \code{decode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeSeq2SeqDecoder$decode(emissions, T, N)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{emissions}}{a emissions}

\item{\code{T}}{a T}

\item{\code{N}}{a N}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeSeq2SeqDecoder-n_hypothesis"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeSeq2SeqDecoder-n_hypothesis}{}}}
\subsection{Method \code{n_hypothesis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeSeq2SeqDecoder$n_hypothesis()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
int
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeSeq2SeqDecoder-prune"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeSeq2SeqDecoder-prune}{}}}
\subsection{Method \code{prune()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeSeq2SeqDecoder$prune(lookBack = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lookBack}}{a lookBack}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeSeq2SeqDecoder-n_decoded_frames_in_buffer"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeSeq2SeqDecoder-n_decoded_frames_in_buffer}{}}}
\subsection{Method \code{n_decoded_frames_in_buffer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeSeq2SeqDecoder$n_decoded_frames_in_buffer()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
int
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeSeq2SeqDecoder-get_best_hypothesis"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeSeq2SeqDecoder-get_best_hypothesis}{}}}
\subsection{Method \code{get_best_hypothesis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeSeq2SeqDecoder$get_best_hypothesis(lookBack = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lookBack}}{a lookBack}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
DecodeResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeSeq2SeqDecoder-get_all_final_hypothesis"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeSeq2SeqDecoder-get_all_final_hypothesis}{}}}
\subsection{Method \code{get_all_final_hypothesis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeSeq2SeqDecoder$get_all_final_hypothesis()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of DecodeResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeSeq2SeqDecoder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeSeq2SeqDecoder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeSeq2SeqDecoder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
