% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon_seq2seq_decoder.R
\name{LexiconSeq2SeqDecoder}
\alias{LexiconSeq2SeqDecoder}
\title{LexiconSeq2SeqDecoder}
\description{
Decoder implements a beam seach decoder that finds the token transcription.
W maximizing:
}
\details{
AM(W) + lmWeight_ * log(P_{lm}(W)) + eosScore_ * |W_last == EOS|

where P_{lm}(W) is the language model score. The transcription W is 
constrained by a lexicon. The language model may operate at word-level 
(isLmToken = FALSE) or token-level (isLmToken = TRUE).


TODO: Doesn't support online decoding now.
}
\section{Super class}{
\code{\link[flashlighttext:Decoder]{flashlighttext::Decoder}} -> \code{LexiconSeq2SeqDecoder}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ptr}}{set and get the pointer to a LexiconSeq2SeqDecoder instance.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LexiconSeq2SeqDecoder-new}{\code{LexiconSeq2SeqDecoder$new()}}
\item \href{#method-LexiconSeq2SeqDecoder-decode_begin}{\code{LexiconSeq2SeqDecoder$decode_begin()}}
\item \href{#method-LexiconSeq2SeqDecoder-decode_step}{\code{LexiconSeq2SeqDecoder$decode_step()}}
\item \href{#method-LexiconSeq2SeqDecoder-decode_end}{\code{LexiconSeq2SeqDecoder$decode_end()}}
\item \href{#method-LexiconSeq2SeqDecoder-decode}{\code{LexiconSeq2SeqDecoder$decode()}}
\item \href{#method-LexiconSeq2SeqDecoder-n_hypothesis}{\code{LexiconSeq2SeqDecoder$n_hypothesis()}}
\item \href{#method-LexiconSeq2SeqDecoder-prune}{\code{LexiconSeq2SeqDecoder$prune()}}
\item \href{#method-LexiconSeq2SeqDecoder-n_decoded_frames_in_buffer}{\code{LexiconSeq2SeqDecoder$n_decoded_frames_in_buffer()}}
\item \href{#method-LexiconSeq2SeqDecoder-get_best_hypothesis}{\code{LexiconSeq2SeqDecoder$get_best_hypothesis()}}
\item \href{#method-LexiconSeq2SeqDecoder-get_all_final_hypothesis}{\code{LexiconSeq2SeqDecoder$get_all_final_hypothesis()}}
\item \href{#method-LexiconSeq2SeqDecoder-clone}{\code{LexiconSeq2SeqDecoder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconSeq2SeqDecoder-new"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconSeq2SeqDecoder-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconSeq2SeqDecoder$new(
  options,
  trie,
  lm,
  eos,
  emitting_model_update_func,
  max_output_length,
  transitions,
  is_lm_token
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{options}}{a options}

\item{\code{trie}}{a trie}

\item{\code{lm}}{a lm}

\item{\code{eos}}{a eos}

\item{\code{emitting_model_update_func}}{a emittingModelUpdateFunc}

\item{\code{max_output_length}}{a maxOutputLength}

\item{\code{transitions}}{a transitions}

\item{\code{is_lm_token}}{a is_lm_token}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
LexiconSeq2SeqDecoder
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconSeq2SeqDecoder-decode_begin"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconSeq2SeqDecoder-decode_begin}{}}}
\subsection{Method \code{decode_begin()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconSeq2SeqDecoder$decode_begin()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconSeq2SeqDecoder-decode_step"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconSeq2SeqDecoder-decode_step}{}}}
\subsection{Method \code{decode_step()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconSeq2SeqDecoder$decode_step(emissions, T, N)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{emissions}}{a emissions}

\item{\code{T}}{a T}

\item{\code{N}}{a N}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconSeq2SeqDecoder-decode_end"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconSeq2SeqDecoder-decode_end}{}}}
\subsection{Method \code{decode_end()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconSeq2SeqDecoder$decode_end()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconSeq2SeqDecoder-decode"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconSeq2SeqDecoder-decode}{}}}
\subsection{Method \code{decode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconSeq2SeqDecoder$decode(emissions, T, N)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{emissions}}{a emissions}

\item{\code{T}}{a T}

\item{\code{N}}{a N}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconSeq2SeqDecoder-n_hypothesis"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconSeq2SeqDecoder-n_hypothesis}{}}}
\subsection{Method \code{n_hypothesis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconSeq2SeqDecoder$n_hypothesis()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
int
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconSeq2SeqDecoder-prune"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconSeq2SeqDecoder-prune}{}}}
\subsection{Method \code{prune()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconSeq2SeqDecoder$prune(lookBack = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lookBack}}{a lookBack}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconSeq2SeqDecoder-n_decoded_frames_in_buffer"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconSeq2SeqDecoder-n_decoded_frames_in_buffer}{}}}
\subsection{Method \code{n_decoded_frames_in_buffer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconSeq2SeqDecoder$n_decoded_frames_in_buffer()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
int
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconSeq2SeqDecoder-get_best_hypothesis"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconSeq2SeqDecoder-get_best_hypothesis}{}}}
\subsection{Method \code{get_best_hypothesis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconSeq2SeqDecoder$get_best_hypothesis(lookBack = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lookBack}}{a lookBack}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
DecodeResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconSeq2SeqDecoder-get_all_final_hypothesis"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconSeq2SeqDecoder-get_all_final_hypothesis}{}}}
\subsection{Method \code{get_all_final_hypothesis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconSeq2SeqDecoder$get_all_final_hypothesis()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of DecodeResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconSeq2SeqDecoder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconSeq2SeqDecoder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconSeq2SeqDecoder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
