% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon_seq2seq_decoder.R
\name{LexiconSeq2SeqDecoderOptions}
\alias{LexiconSeq2SeqDecoderOptions}
\title{LexiconSeq2SeqDecoderOptions}
\value{
LexiconSeq2SeqDecoderOptions
}
\description{
LexiconSeq2SeqDecoderOptions

LexiconSeq2SeqDecoderOptions
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{beam_size}}{Maximum number of hypothesis we hold after each step}

\item{\code{beam_size_token}}{Maximum number of tokens we consider at each step}

\item{\code{beam_threshold}}{Threshold to prune hypothesis}

\item{\code{lm_weight}}{Weight of lm}

\item{\code{word_score}}{Word insertion score}

\item{\code{eos_score}}{Score for inserting an EOS}

\item{\code{log_add}}{If or not use logadd when merging hypothesis}

\item{\code{ptr}}{set and get the pointer to a LexiconSeq2SeqDecoderOptions instance.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LexiconSeq2SeqDecoderOptions-new}{\code{LexiconSeq2SeqDecoderOptions$new()}}
\item \href{#method-LexiconSeq2SeqDecoderOptions-clone}{\code{LexiconSeq2SeqDecoderOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconSeq2SeqDecoderOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconSeq2SeqDecoderOptions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconSeq2SeqDecoderOptions$new(
  beam_size,
  beam_size_token,
  beam_threshold,
  lm_weight,
  word_score,
  eos_score,
  log_add
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{beam_size}}{Maximum number of hypothesis we hold after each step}

\item{\code{beam_size_token}}{Maximum number of tokens we consider at each step}

\item{\code{beam_threshold}}{Threshold to prune hypothesis}

\item{\code{lm_weight}}{Weight of lm}

\item{\code{word_score}}{Word insertion score}

\item{\code{eos_score}}{Score for inserting an EOS}

\item{\code{log_add}}{If or not use logadd when merging hypothesis}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconSeq2SeqDecoderOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconSeq2SeqDecoderOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconSeq2SeqDecoderOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
