% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{LM}
\alias{LM}
\title{LM}
\description{
LM

LM
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ptr}}{get or set a pointer to a (*)LMWrapper.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LM-new}{\code{LM$new()}}
\item \href{#method-LM-start}{\code{LM$start()}}
\item \href{#method-LM-score}{\code{LM$score()}}
\item \href{#method-LM-finish}{\code{LM$finish()}}
\item \href{#method-LM-clone}{\code{LM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LM-new"></a>}}
\if{latex}{\out{\hypertarget{method-LM-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LM$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LM-start"></a>}}
\if{latex}{\out{\hypertarget{method-LM-start}{}}}
\subsection{Method \code{start()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LM$start(startWithNothing)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{startWithNothing}}{a logical value indicating whether the state 
should be initialized with a start-of-sequence marker.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An LMState object representing the current state.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LM-score"></a>}}
\if{latex}{\out{\hypertarget{method-LM-score}{}}}
\subsection{Method \code{score()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LM$score(state, usrTokenIdx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{state}}{An LMState object representing the current state.}

\item{\code{usrTokenIdx}}{An integer representing the index of the current token 
in the user-defined token list.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing the externalpointer to the updated LMState object
and a numeric value representing the score.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LM-finish"></a>}}
\if{latex}{\out{\hypertarget{method-LM-finish}{}}}
\subsection{Method \code{finish()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LM$finish(state)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{state}}{An LMState object representing the current state.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing the updated LMState object and a numeric value 
representing the final score.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LM-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
