% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmstate.R
\name{LMState}
\alias{LMState}
\title{LMState}
\description{
LMState

LMState
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ptr}}{get or set a pointer to a LMStateWrapper.}

\item{\code{children}}{a list of pairs (index, LMStatePtr)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LMState-new}{\code{LMState$new()}}
\item \href{#method-LMState-child}{\code{LMState$child()}}
\item \href{#method-LMState-compare}{\code{LMState$compare()}}
\item \href{#method-LMState-clone}{\code{LMState$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LMState-new"></a>}}
\if{latex}{\out{\hypertarget{method-LMState-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LMState$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An instance of the LMState class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LMState-child"></a>}}
\if{latex}{\out{\hypertarget{method-LMState-child}{}}}
\subsection{Method \code{child()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LMState$child(usr_index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{usr_index}}{a usr_index}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
externalpointer to a LMStateWrapper
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LMState-compare"></a>}}
\if{latex}{\out{\hypertarget{method-LMState-compare}{}}}
\subsection{Method \code{compare()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LMState$compare(state)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{state}}{an externalpointer to a LMState to compare with.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
int
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LMState-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LMState-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LMState$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
