% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{load_words}
\alias{load_words}
\title{load_words}
\usage{
load_words(filename, maxWords = -1)
}
\arguments{
\item{filename}{a string}

\item{maxWords}{an int}
}
\value{
LexiconMap
}
\description{
load_words
}
