% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{pack_replabels}
\alias{pack_replabels}
\title{pack_replabels}
\usage{
pack_replabels(tokens, dict, max_reps)
}
\arguments{
\item{tokens}{int vector}

\item{dict}{Dictionary instance}

\item{max_reps}{atomic int}
}
\value{
int vector
}
\description{
pack_replabels
}
