% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trie.R
\docType{data}
\name{SmearingModes}
\alias{SmearingModes}
\title{SmearingModes}
\format{
## `SmearingModes`
A list with 3 items:
\describe{
  \item{NONE}{Default = 0}
  \item{MAX}{Default = 1}
  \item{LOGADD}{Default = 2}
}
}
\usage{
SmearingModes
}
\description{
A enum (list) of semearing modes.
}
\keyword{datasets}
