% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trie.R
\name{TrieNode}
\alias{TrieNode}
\title{TrieNode}
\description{
TrieNode is the trie node structure in Trie.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ptr}}{has a single parameter new_ptr that accepts a new pointer to a TrieNode.
It returns invisible(NULL)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TrieNode-new}{\code{TrieNode$new()}}
\item \href{#method-TrieNode-max_score}{\code{TrieNode$max_score()}}
\item \href{#method-TrieNode-clone}{\code{TrieNode$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrieNode-new"></a>}}
\if{latex}{\out{\hypertarget{method-TrieNode-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrieNode$new(idx = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{idx}}{an integer. Node index}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TrieNode
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrieNode-max_score"></a>}}
\if{latex}{\out{\hypertarget{method-TrieNode-max_score}{}}}
\subsection{Method \code{max_score()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrieNode$max_score()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the node max score
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrieNode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TrieNode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrieNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
