% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{unpack_replabels}
\alias{unpack_replabels}
\title{unpack_replabels}
\usage{
unpack_replabels(tokens, dict, max_reps)
}
\arguments{
\item{tokens}{int vector}

\item{dict}{Dictionary instance}

\item{max_reps}{atomic int}
}
\value{
int vector
}
\description{
unpack_replabels
}
