% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zerolm.R
\name{ZeroLM}
\alias{ZeroLM}
\title{ZeroLM}
\description{
ZeroLM

ZeroLM
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ptr}}{has a single parameter new_ptr that accepts a new pointer to a ZeroLMWrapper.
It returns invisible(NULL)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ZeroLM-new}{\code{ZeroLM$new()}}
\item \href{#method-ZeroLM-start}{\code{ZeroLM$start()}}
\item \href{#method-ZeroLM-score}{\code{ZeroLM$score()}}
\item \href{#method-ZeroLM-finish}{\code{ZeroLM$finish()}}
\item \href{#method-ZeroLM-clone}{\code{ZeroLM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZeroLM-new"></a>}}
\if{latex}{\out{\hypertarget{method-ZeroLM-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZeroLM$new()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{a path}

\item{\code{usrTknDict}}{a usrTknDict}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ZeroLM
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZeroLM-start"></a>}}
\if{latex}{\out{\hypertarget{method-ZeroLM-start}{}}}
\subsection{Method \code{start()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZeroLM$start(startWithNothing)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{startWithNothing}}{a boolean}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
LMState
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZeroLM-score"></a>}}
\if{latex}{\out{\hypertarget{method-ZeroLM-score}{}}}
\subsection{Method \code{score()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZeroLM$score(state, usrTokenIdx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{state}}{a state}

\item{\code{usrTokenIdx}}{a usrTokenIdx}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list(LMState, numeric)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZeroLM-finish"></a>}}
\if{latex}{\out{\hypertarget{method-ZeroLM-finish}{}}}
\subsection{Method \code{finish()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZeroLM$finish(state)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{state}}{a state}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list(LMState, numeric)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZeroLM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ZeroLM-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZeroLM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
