% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoder.R
\name{Decoder}
\alias{Decoder}
\title{Decoder}
\description{
Decoder

Decoder
}
\examples{

library(flashlighttext)

decoder <- Decoder$new()
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ptr}}{set and get the pointer to a Decoder instance.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Decoder-new}{\code{Decoder$new()}}
\item \href{#method-Decoder-decodeBegin}{\code{Decoder$decodeBegin()}}
\item \href{#method-Decoder-decodeStep}{\code{Decoder$decodeStep()}}
\item \href{#method-Decoder-decodeEnd}{\code{Decoder$decodeEnd()}}
\item \href{#method-Decoder-decode}{\code{Decoder$decode()}}
\item \href{#method-Decoder-prune}{\code{Decoder$prune()}}
\item \href{#method-Decoder-nDecodedFramesInBuffer}{\code{Decoder$nDecodedFramesInBuffer()}}
\item \href{#method-Decoder-getBestHypothesis}{\code{Decoder$getBestHypothesis()}}
\item \href{#method-Decoder-getAllFinalHypothesis}{\code{Decoder$getAllFinalHypothesis()}}
\item \href{#method-Decoder-clone}{\code{Decoder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Decoder-new"></a>}}
\if{latex}{\out{\hypertarget{method-Decoder-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Decoder$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Decoder-decodeBegin"></a>}}
\if{latex}{\out{\hypertarget{method-Decoder-decodeBegin}{}}}
\subsection{Method \code{decodeBegin()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Decoder$decodeBegin()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Decoder-decodeStep"></a>}}
\if{latex}{\out{\hypertarget{method-Decoder-decodeStep}{}}}
\subsection{Method \code{decodeStep()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Decoder$decodeStep(emissions, T, N)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{emissions}}{a emissions}

\item{\code{T}}{a T}

\item{\code{N}}{a N}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Decoder-decodeEnd"></a>}}
\if{latex}{\out{\hypertarget{method-Decoder-decodeEnd}{}}}
\subsection{Method \code{decodeEnd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Decoder$decodeEnd()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Decoder-decode"></a>}}
\if{latex}{\out{\hypertarget{method-Decoder-decode}{}}}
\subsection{Method \code{decode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Decoder$decode(emissions, T, N)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{emissions}}{a emissions}

\item{\code{T}}{a T}

\item{\code{N}}{a N}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Decoder-prune"></a>}}
\if{latex}{\out{\hypertarget{method-Decoder-prune}{}}}
\subsection{Method \code{prune()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Decoder$prune(lookBack = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lookBack}}{a lookBack}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Decoder-nDecodedFramesInBuffer"></a>}}
\if{latex}{\out{\hypertarget{method-Decoder-nDecodedFramesInBuffer}{}}}
\subsection{Method \code{nDecodedFramesInBuffer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Decoder$nDecodedFramesInBuffer()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Decoder-getBestHypothesis"></a>}}
\if{latex}{\out{\hypertarget{method-Decoder-getBestHypothesis}{}}}
\subsection{Method \code{getBestHypothesis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Decoder$getBestHypothesis(lookBack = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lookBack}}{a lookBack}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Decoder-getAllFinalHypothesis"></a>}}
\if{latex}{\out{\hypertarget{method-Decoder-getAllFinalHypothesis}{}}}
\subsection{Method \code{getAllFinalHypothesis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Decoder$getAllFinalHypothesis()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Decoder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Decoder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Decoder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
