% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon_decoder.R
\name{LexiconDecoder}
\alias{LexiconDecoder}
\title{LexiconDecoder}
\description{
LexiconDecoder

LexiconDecoder
}
\section{Super class}{
\code{\link[flashlighttext:Decoder]{flashlighttext::Decoder}} -> \code{LexiconDecoder}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ptr}}{set and get the pointer to a LexiconDecoder instance.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LexiconDecoder-new}{\code{LexiconDecoder$new()}}
\item \href{#method-LexiconDecoder-decode_begin}{\code{LexiconDecoder$decode_begin()}}
\item \href{#method-LexiconDecoder-decode_step}{\code{LexiconDecoder$decode_step()}}
\item \href{#method-LexiconDecoder-decode_end}{\code{LexiconDecoder$decode_end()}}
\item \href{#method-LexiconDecoder-decode}{\code{LexiconDecoder$decode()}}
\item \href{#method-LexiconDecoder-n_hypothesis}{\code{LexiconDecoder$n_hypothesis()}}
\item \href{#method-LexiconDecoder-prune}{\code{LexiconDecoder$prune()}}
\item \href{#method-LexiconDecoder-n_decoded_frames_in_buffer}{\code{LexiconDecoder$n_decoded_frames_in_buffer()}}
\item \href{#method-LexiconDecoder-get_best_hypothesis}{\code{LexiconDecoder$get_best_hypothesis()}}
\item \href{#method-LexiconDecoder-get_all_final_hypothesis}{\code{LexiconDecoder$get_all_final_hypothesis()}}
\item \href{#method-LexiconDecoder-clone}{\code{LexiconDecoder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconDecoder-new"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconDecoder-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconDecoder$new(
  options,
  trie,
  lm,
  sil_token_idx,
  blank_token_idx,
  unk_token_idx,
  transitions,
  is_lm_token
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{options}}{a options}

\item{\code{trie}}{a trie}

\item{\code{lm}}{a lm}

\item{\code{sil_token_idx}}{a sil_token_idx}

\item{\code{blank_token_idx}}{a blank_token_idx}

\item{\code{unk_token_idx}}{a unk_token_idx}

\item{\code{transitions}}{a transitions}

\item{\code{is_lm_token}}{a is_lm_token}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
LexiconDecoder
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconDecoder-decode_begin"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconDecoder-decode_begin}{}}}
\subsection{Method \code{decode_begin()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconDecoder$decode_begin()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconDecoder-decode_step"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconDecoder-decode_step}{}}}
\subsection{Method \code{decode_step()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconDecoder$decode_step(emissions, T, N)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{emissions}}{a emissions}

\item{\code{T}}{a T}

\item{\code{N}}{a N}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconDecoder-decode_end"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconDecoder-decode_end}{}}}
\subsection{Method \code{decode_end()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconDecoder$decode_end()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconDecoder-decode"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconDecoder-decode}{}}}
\subsection{Method \code{decode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconDecoder$decode(emissions, T, N)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{emissions}}{a emissions}

\item{\code{T}}{a T}

\item{\code{N}}{a N}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconDecoder-n_hypothesis"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconDecoder-n_hypothesis}{}}}
\subsection{Method \code{n_hypothesis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconDecoder$n_hypothesis()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
int
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconDecoder-prune"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconDecoder-prune}{}}}
\subsection{Method \code{prune()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconDecoder$prune(lookBack = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lookBack}}{a lookBack}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible(NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconDecoder-n_decoded_frames_in_buffer"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconDecoder-n_decoded_frames_in_buffer}{}}}
\subsection{Method \code{n_decoded_frames_in_buffer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconDecoder$n_decoded_frames_in_buffer()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
int
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconDecoder-get_best_hypothesis"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconDecoder-get_best_hypothesis}{}}}
\subsection{Method \code{get_best_hypothesis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconDecoder$get_best_hypothesis(lookBack = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lookBack}}{a lookBack}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
DecodeResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconDecoder-get_all_final_hypothesis"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconDecoder-get_all_final_hypothesis}{}}}
\subsection{Method \code{get_all_final_hypothesis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconDecoder$get_all_final_hypothesis()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of DecodeResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconDecoder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconDecoder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconDecoder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
