% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon_free_decoder.R
\name{LexiconFreeDecoder}
\alias{LexiconFreeDecoder}
\title{LexiconFreeDecoder}
\description{
LexiconFreeDecoder

LexiconFreeDecoder
}
\section{Super class}{
\code{\link[flashlighttext:Decoder]{flashlighttext::Decoder}} -> \code{LexiconFreeDecoder}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LexiconFreeDecoder-new}{\code{LexiconFreeDecoder$new()}}
\item \href{#method-LexiconFreeDecoder-n_hypothesis}{\code{LexiconFreeDecoder$n_hypothesis()}}
\item \href{#method-LexiconFreeDecoder-get_lm_ptr}{\code{LexiconFreeDecoder$get_lm_ptr()}}
\item \href{#method-LexiconFreeDecoder-get_sil_idx}{\code{LexiconFreeDecoder$get_sil_idx()}}
\item \href{#method-LexiconFreeDecoder-get_blank_idx}{\code{LexiconFreeDecoder$get_blank_idx()}}
\item \href{#method-LexiconFreeDecoder-get_options}{\code{LexiconFreeDecoder$get_options()}}
\item \href{#method-LexiconFreeDecoder-get_transitions}{\code{LexiconFreeDecoder$get_transitions()}}
\item \href{#method-LexiconFreeDecoder-clone}{\code{LexiconFreeDecoder$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="flashlighttext" data-topic="Decoder" data-id="decode"><a href='../../flashlighttext/html/Decoder.html#method-Decoder-decode'><code>flashlighttext::Decoder$decode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="flashlighttext" data-topic="Decoder" data-id="get_all_final_hypothesis"><a href='../../flashlighttext/html/Decoder.html#method-Decoder-get_all_final_hypothesis'><code>flashlighttext::Decoder$get_all_final_hypothesis()</code></a></span></li>
<li><span class="pkg-link" data-pkg="flashlighttext" data-topic="Decoder" data-id="get_best_hypothesis"><a href='../../flashlighttext/html/Decoder.html#method-Decoder-get_best_hypothesis'><code>flashlighttext::Decoder$get_best_hypothesis()</code></a></span></li>
<li><span class="pkg-link" data-pkg="flashlighttext" data-topic="Decoder" data-id="n_decoded_frames_in_buffer"><a href='../../flashlighttext/html/Decoder.html#method-Decoder-n_decoded_frames_in_buffer'><code>flashlighttext::Decoder$n_decoded_frames_in_buffer()</code></a></span></li>
<li><span class="pkg-link" data-pkg="flashlighttext" data-topic="Decoder" data-id="prune"><a href='../../flashlighttext/html/Decoder.html#method-Decoder-prune'><code>flashlighttext::Decoder$prune()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoder-new"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoder-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoder$new(
  options,
  lm,
  sil_token_idx,
  blank_token_idx,
  transitions
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{options}}{a options}

\item{\code{lm}}{a lm}

\item{\code{sil_token_idx}}{a sil_token_idx}

\item{\code{blank_token_idx}}{a blank_token_idx}

\item{\code{transitions}}{a transitions}

\item{\code{is_lm_token}}{a is_lm_token}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
LexiconFreeDecoder
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoder-n_hypothesis"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoder-n_hypothesis}{}}}
\subsection{Method \code{n_hypothesis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoder$n_hypothesis()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
int
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoder-get_lm_ptr"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoder-get_lm_ptr}{}}}
\subsection{Method \code{get_lm_ptr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoder$get_lm_ptr()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a externalpointer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoder-get_sil_idx"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoder-get_sil_idx}{}}}
\subsection{Method \code{get_sil_idx()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoder$get_sil_idx()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of DecodeResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoder-get_blank_idx"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoder-get_blank_idx}{}}}
\subsection{Method \code{get_blank_idx()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoder$get_blank_idx()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of DecodeResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoder-get_options"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoder-get_options}{}}}
\subsection{Method \code{get_options()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoder$get_options()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of DecodeResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoder-get_transitions"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoder-get_transitions}{}}}
\subsection{Method \code{get_transitions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoder$get_transitions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of DecodeResult
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeDecoder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeDecoder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeDecoder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
