% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon_free_seq2seq_decoder.R
\name{LexiconFreeSeq2SeqDecoder}
\alias{LexiconFreeSeq2SeqDecoder}
\title{LexiconFreeSeq2SeqDecoder}
\description{
LexiconFreeSeq2SeqDecoder

LexiconFreeSeq2SeqDecoder
}
\section{Super class}{
\code{\link[flashlighttext:Decoder]{flashlighttext::Decoder}} -> \code{LexiconFreeSeq2SeqDecoder}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LexiconFreeSeq2SeqDecoder-new}{\code{LexiconFreeSeq2SeqDecoder$new()}}
\item \href{#method-LexiconFreeSeq2SeqDecoder-n_hypothesis}{\code{LexiconFreeSeq2SeqDecoder$n_hypothesis()}}
\item \href{#method-LexiconFreeSeq2SeqDecoder-clone}{\code{LexiconFreeSeq2SeqDecoder$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="flashlighttext" data-topic="Decoder" data-id="decode"><a href='../../flashlighttext/html/Decoder.html#method-Decoder-decode'><code>flashlighttext::Decoder$decode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="flashlighttext" data-topic="Decoder" data-id="get_all_final_hypothesis"><a href='../../flashlighttext/html/Decoder.html#method-Decoder-get_all_final_hypothesis'><code>flashlighttext::Decoder$get_all_final_hypothesis()</code></a></span></li>
<li><span class="pkg-link" data-pkg="flashlighttext" data-topic="Decoder" data-id="get_best_hypothesis"><a href='../../flashlighttext/html/Decoder.html#method-Decoder-get_best_hypothesis'><code>flashlighttext::Decoder$get_best_hypothesis()</code></a></span></li>
<li><span class="pkg-link" data-pkg="flashlighttext" data-topic="Decoder" data-id="n_decoded_frames_in_buffer"><a href='../../flashlighttext/html/Decoder.html#method-Decoder-n_decoded_frames_in_buffer'><code>flashlighttext::Decoder$n_decoded_frames_in_buffer()</code></a></span></li>
<li><span class="pkg-link" data-pkg="flashlighttext" data-topic="Decoder" data-id="prune"><a href='../../flashlighttext/html/Decoder.html#method-Decoder-prune'><code>flashlighttext::Decoder$prune()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeSeq2SeqDecoder-new"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeSeq2SeqDecoder-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeSeq2SeqDecoder$new(
  options,
  lm,
  eos_token_idx,
  emitting_model_update_func,
  max_output_length
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{options}}{an options}

\item{\code{lm}}{a lm}

\item{\code{eos_token_idx}}{an eos_token_idx}

\item{\code{emitting_model_update_func}}{an emitting_model_update_func}

\item{\code{max_output_length}}{a max_output_length}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
LexiconFreeSeq2SeqDecoder
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeSeq2SeqDecoder-n_hypothesis"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeSeq2SeqDecoder-n_hypothesis}{}}}
\subsection{Method \code{n_hypothesis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeSeq2SeqDecoder$n_hypothesis()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
int
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconFreeSeq2SeqDecoder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconFreeSeq2SeqDecoder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconFreeSeq2SeqDecoder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
