% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon_seq2seq_decoder.R
\name{LexiconSeq2SeqDecoderOptions}
\alias{LexiconSeq2SeqDecoderOptions}
\title{LexiconSeq2SeqDecoderOptions}
\value{
LexiconSeq2SeqDecoderOptions
}
\description{
LexiconSeq2SeqDecoderOptions

LexiconSeq2SeqDecoderOptions
}
\section{Super class}{
\code{\link[flashlighttext:Decoder]{flashlighttext::Decoder}} -> \code{LexiconSeq2SeqDecoderOptions}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{beam_size}}{Maximum number of hypothesis we hold after each step}

\item{\code{beam_size_token}}{Maximum number of tokens we consider at each step}

\item{\code{beam_threshold}}{Threshold to prune hypothesis}

\item{\code{lm_weight}}{Weight of lm}

\item{\code{word_score}}{Word insertion score}

\item{\code{eos_score}}{Score for inserting an EOS}

\item{\code{log_add}}{If or not use logadd when merging hypothesis}

\item{\code{ptr}}{set and get the pointer to a LexiconSeq2SeqDecoderOptions instance.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LexiconSeq2SeqDecoderOptions-new}{\code{LexiconSeq2SeqDecoderOptions$new()}}
\item \href{#method-LexiconSeq2SeqDecoderOptions-clone}{\code{LexiconSeq2SeqDecoderOptions$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="flashlighttext" data-topic="Decoder" data-id="decode"><a href='../../flashlighttext/html/Decoder.html#method-Decoder-decode'><code>flashlighttext::Decoder$decode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="flashlighttext" data-topic="Decoder" data-id="get_all_final_hypothesis"><a href='../../flashlighttext/html/Decoder.html#method-Decoder-get_all_final_hypothesis'><code>flashlighttext::Decoder$get_all_final_hypothesis()</code></a></span></li>
<li><span class="pkg-link" data-pkg="flashlighttext" data-topic="Decoder" data-id="get_best_hypothesis"><a href='../../flashlighttext/html/Decoder.html#method-Decoder-get_best_hypothesis'><code>flashlighttext::Decoder$get_best_hypothesis()</code></a></span></li>
<li><span class="pkg-link" data-pkg="flashlighttext" data-topic="Decoder" data-id="n_decoded_frames_in_buffer"><a href='../../flashlighttext/html/Decoder.html#method-Decoder-n_decoded_frames_in_buffer'><code>flashlighttext::Decoder$n_decoded_frames_in_buffer()</code></a></span></li>
<li><span class="pkg-link" data-pkg="flashlighttext" data-topic="Decoder" data-id="prune"><a href='../../flashlighttext/html/Decoder.html#method-Decoder-prune'><code>flashlighttext::Decoder$prune()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconSeq2SeqDecoderOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconSeq2SeqDecoderOptions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconSeq2SeqDecoderOptions$new(
  beam_size,
  beam_size_token,
  beam_threshold,
  lm_weight,
  word_score,
  eos_score,
  log_add
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{beam_size}}{an integer. Maximum number of hypothesis we hold after each step}

\item{\code{beam_size_token}}{an integer. Maximum number of tokens we consider at each step}

\item{\code{beam_threshold}}{a numeric. Threshold to prune hypothesis}

\item{\code{lm_weight}}{a numeric. Weight of lm}

\item{\code{word_score}}{a numeric. Word insertion score}

\item{\code{eos_score}}{a numeric. Score for inserting an EOS}

\item{\code{log_add}}{a logical. If or not use logadd when merging hypothesis}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LexiconSeq2SeqDecoderOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LexiconSeq2SeqDecoderOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LexiconSeq2SeqDecoderOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
