% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zerolm.R
\name{ZeroLM}
\alias{ZeroLM}
\title{ZeroLM}
\description{
ZeroLM is a dummy language model class, which mimics the behaviour of a
uni-gram language model but always returns 0 as score.
}
\section{Super class}{
\code{\link[flashlighttext:LM]{flashlighttext::LM}} -> \code{ZeroLM}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ZeroLM-new}{\code{ZeroLM$new()}}
\item \href{#method-ZeroLM-clone}{\code{ZeroLM$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="flashlighttext" data-topic="LM" data-id="finish"><a href='../../flashlighttext/html/LM.html#method-LM-finish'><code>flashlighttext::LM$finish()</code></a></span></li>
<li><span class="pkg-link" data-pkg="flashlighttext" data-topic="LM" data-id="score"><a href='../../flashlighttext/html/LM.html#method-LM-score'><code>flashlighttext::LM$score()</code></a></span></li>
<li><span class="pkg-link" data-pkg="flashlighttext" data-topic="LM" data-id="start"><a href='../../flashlighttext/html/LM.html#method-LM-start'><code>flashlighttext::LM$start()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZeroLM-new"></a>}}
\if{latex}{\out{\hypertarget{method-ZeroLM-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZeroLM$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ZeroLM
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZeroLM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ZeroLM-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZeroLM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
